Address SA reviewer test cleanup
